refactor: move starters to seperate validation file
