+++ sage/utils/validator.py
@@ -5,7 +5,6 @@ from logging import getLevelName
 from typing import List, Literal, Optional
 
 from anyio import Path
-from chainlit import Starter
 from croniter import croniter
 from pydantic import (
     BaseModel,
@@ -231,6 +230,9 @@ class Core(BaseModel):
     logging_level: str | int = "INFO"
     user_agent: str = "codesage.ai"
 
+    starters_path: Optional[str] = None
+    agents_dir: Optional[str] = None
+
     @field_validator("logging_level")
     @classmethod
     def set_logging_level(cls, v: str | int):
@@ -295,29 +297,6 @@ class LLMConfig(BaseModel):
     """The configuration for LLM models"""
 
     model: str
-
-
-class StarterConfig(BaseModel, Starter):
-    label: str
-    message: str
-    icon: Optional[str] = None
-
-    @model_validator(mode="before")
-    @classmethod
-    def combine_message_and_source(cls, values: dict) -> dict:
-        """Combine the message and source to create a new message-source data"""
-        message: str = values.get("message")
-        if not message:
-            raise ConfigException("The message field is missing")
-        source: str = values.get("source", "none").strip()
-        values["message"] = f"{message} %{source}%".strip()
-        return values
-
-
-class Starters(BaseModel):
-    """Starters config model"""
-
-    starters: List[StarterConfig] = Field(default=[])
 
 
 class Config(BaseModel):
