+++ tests/unit_tests/utils/test_validator.py
@@ -24,8 +24,6 @@ from sage.utils.validator import (
     ReRankerConfig,
     Source,
     SourceData,
-    StarterConfig,
-    Starters,
     UploadConfig,
     Web,
 )
@@ -373,17 +371,24 @@ def test_core_default_values():
     assert core.data_dir == Path.home() / ".sage"
     assert core.logging_level == "INFO"
     assert core.user_agent == "codesage.ai"
+    assert core.starters_path == None
+    assert core.agents_dir == None
 
 
 def test_core_custom_values():
     custom_data_dir = "/custom/path"
     core = Core(
-        data_dir=custom_data_dir, logging_level="DEBUG", user_agent="custom_agent"
+        data_dir=custom_data_dir,
+        logging_level="DEBUG",
+        user_agent="custom_agent",
+        starters_path="/fake/starters.yaml",
+        agents_dir="/fake/agents",
     )
     assert str(core.data_dir) == custom_data_dir
     assert core.logging_level == getLevelName("DEBUG")
     assert core.user_agent == "custom_agent"
-
+    assert core.starters_path == "/fake/starters.yaml"
+    assert core.agents_dir == "/fake/agents"
 
 def test_core_logging_level_validation():
     core = Core(logging_level="WARNING")
@@ -587,55 +592,3 @@ def test_config_default_optional_fields(setup_env_vars):
     assert isinstance(config.core, Core)
     assert isinstance(config.upload, UploadConfig)
     assert config.reranker is None
-
-
-###########################################################################
-################ Unit Tests for the StartersConfig ########################
-
-
-@pytest.fixture
-def starter_data():
-    return {
-        "label": "Daily Summary",
-        "message": "Show me the daily summary",
-        "icon": "https://example.com/icon.png",
-    }
-
-
-def test_valid_starter_config(starter_data):
-    starter = StarterConfig(**starter_data)
-    assert starter.label == starter_data["label"]
-    assert starter.message == starter_data["message"] + " %none%"
-    assert starter.icon == starter_data["icon"]
-
-
-def test_starter_config_missing_fields():
-    with pytest.raises(ConfigException):
-        StarterConfig(label="Weather Report")
-
-
-def test_starter_config_with_source():
-    starter = StarterConfig(
-        label="Weather Report", message="What is the weather?", source="yahoo"
-    )
-    assert starter.message == "What is the weather? %yahoo%"
-
-
-def test_valid_starter_config_with_starters(starter_data):
-    test_starter = StarterConfig(**starter_data)
-    starter_config = Starters(starters=[test_starter])
-    assert len(starter_config.starters) == 1
-    assert starter_config.starters[0].label == starter_data["label"]
-    assert starter_config.starters[0].message == starter_data["message"] + " %none%"
-    assert starter_config.starters[0].icon == starter_data["icon"]
-
-
-def test_starter_config_empty_starters_list():
-    starter_config = Starters(starters=[])
-    assert len(starter_config.starters) == 0
-
-
-def test_starters_invalid_type():
-    # Starters should be a list
-    with pytest.raises(ValidationError):
-        Starters(starters="not a list")
